Add request timing option for access logger
Add the ability to specify a boolean configuration parameter to the
webmachine_access_log_handler module to have it include request
duration reporting as part of its log output. The request durations
are reported in microseconds and this option is disabled by default so
that the current default behavior is maintained.